refactor(platform): extract audio config into config/audio.rs (#2830)
Move audio CLI/config-value parsing, sample-rate validation, and audio
constants into config/audio.rs as per-domain apply_args/apply_config_value
functions. Add shared test_support helpers and relocate audio tests. mod.rs
orchestrators delegate to the audio module. No behavior change.
